Changing unit price for new order row to long from decimal

+++ Webpay.Integration.CSharp/Webpay.Integration.CSharp/Order/Row/Credit/NewCreditOrderRowBuilder.cs
@@ -3,7 +3,7 @@
     public class NewCreditOrderRowBuilder 
     {
         public string Name { get; set; }
-        public decimal UnitPrice { get; set; }
+        public long UnitPrice { get; set; }
         public decimal Quantity { get; set; }
         public decimal VatPercent { get; set; }
         public decimal DiscountAmount { get; set; }
